Removing PodCoverageTests from Dist::Zilla config
Now

    dzil test --author --release

runs without error.

The reason for this removal (which involves replacing the [@Classic] D::Z
plugin with [@Basic] and augmenting it to achieve the same behaviour,
however without PodCoverageTests) was that PodCoverageTests gave a warning
for *every* aliased Given/When/Then for *every* language...  The other
option would have been to maintain a `xt/pod-coverage.t` or similar which
then loads all possible language versions of Given/When/Then from the
`i18n.json` and tells Pod::Coverage to ignore them.  Also one would have to
tell D::Z not to generate `release-pod-coverage.t` etc...  In short, it
looks like the hassle of keeping PodCoverageTests seems higher than simply
removing it.